refactor(maencof): decompose mcp/server/lifecycle into operations/ organ
Move bootSweep.ts and registerShutdown.ts into lifecycle/operations/
(known organ), re-exported via the index.ts barrel — clears the
zero-peer-file warnings and matches the fractal-root principle
(root = barrel + INTENT only, logic in an organ subdir). server.ts is
unaffected (barrel path unchanged); mcpLifecycle.test.ts imports via the
barrel. structure_validate: 0 violations; typecheck clean; suite 1223 passed.
