+++ plugins/maencof/src/__tests__/unit/mcpLifecycle.test.ts
@@ -11,8 +11,7 @@ import { sweepStaleSessions } from '../../core/sessionStore/index.js';
 import { buildDailyDigest } from '../../core/workIndex/index.js';
 import { isMaencofVault } from '../../hooks/shared/isMaencofVault.js';
 import { runVaultCommitter } from '../../hooks/utils/vaultCommitter/operations/runVaultCommitter.js';
-import { bootSweep } from '../../mcp/server/lifecycle/bootSweep.js';
-import { registerShutdown } from '../../mcp/server/lifecycle/registerShutdown.js';
+import { bootSweep, registerShutdown } from '../../mcp/server/lifecycle/index.js';
 
 const calls: string[] = [];
